Update packages in experimental branch with packages from master branch.
This adds the new packages cisco and netflow and makes sure the versions are in line with the one in master. Further changes to the packages in master should be picked one by one into this branch if needed. We need to be especially careful not to pick any chances in which are based on data streams.