Add testing for packages
This adds code which can be run to test the packages against a Kibana instance. What it does it setup the Elastic Stack, runs setup for ingest manager and then installs all the package into Kibana which are served by the registry. This means these are only the most recent packages, not all of them.

In the future, we could only test the packages which are part of the PR instead of all of them.

Below is an example output of the test:

```
--- FAIL: TestSetup (162.99s)
    --- FAIL: TestSetup/install-packages (0.00s)
        --- PASS: TestSetup/install-packages/log-0.1.0 (13.73s)
        --- PASS: TestSetup/install-packages/endpoint-0.1.0 (13.75s)
        --- PASS: TestSetup/install-packages/kafka-0.1.0 (17.67s)
        --- PASS: TestSetup/install-packages/mysql-0.1.0 (19.90s)
        --- PASS: TestSetup/install-packages/netflow-0.1.0 (21.77s)
        --- PASS: TestSetup/install-packages/nginx-0.1.0 (23.64s)
        --- FAIL: TestSetup/install-packages/cisco-0.1.0 (10.95s)
        --- PASS: TestSetup/install-packages/redis-0.1.0 (24.76s)
        --- PASS: TestSetup/install-packages/system-0.1.0 (16.35s)
        --- PASS: TestSetup/install-packages/aws-0.0.3 (35.16s)
FAIL
FAIL    github.com/elastic/package-storage/testing      163.074s
FAIL
```