Add testing for packages (#54)
This adds code which can be run to test the packages against a Kibana instance. What it does it setup the Elastic Stack, runs setup for ingest manager and then installs all the package into Kibana which are served by the registry. This means these are only the most recent packages, not all of them.

In the future, we could only test the packages which are part of the PR instead of all of them.