Remove logs/metrics categories (#97)
The logs and metrics categories do not exist anymore. In https://github.com/elastic/package-registry/pull/534 the supported categories were updated.

The categories are removed for now to allow the registry to remove the legacy support for these categories. The new releases of these packages must contain the new categories.