Remove packages as master is not production anymore (#116)
As part of https://github.com/elastic/package-storage/issues/86 the master branch will not be used anymore. By now, all packages have been moved over to the production branch and it is ready to be deployed under `epr.elastic.co`. All future contributions of packages should go to the production branch or as soon as staging and snapshot deployment are fully available + promotion script, follow the new process.

This PR removes all packages and big junk of the code. Further cleanup will be needed. The goal of this PR is to make sure this registry is not used anymore moving forward.