[production] Improve integration testing script (#345)
This change improves the integration testing script by removing the parallel execution. On of the flaky parts around this script was, that sometimes Kibana was overloaded by the number of requests and returned a 429. The tests were run in parallel as initially installing a package was much slower then it is now. Even though this change means tests take between 60-120 seconds longer overall, I think the benefit of having less flaky tests are worth the change.

Long term, it would be great if the tests are only run on the packages that actually changed instead of all packages.

To clean up the code a bit, `installPackage` was moved into its own method.