[production] Increase cache time for assets
On production, we should use a higher cache time for assets. Especially the static assets should not change. The search API has only a cache time of 1min to show new packages quickly. The category endpoint is 10m as if the counter of a package is not immedititely correct it is not an issue.

Long term, we should increase these values further.