Deprecate rsa2elk package for Symantec
This package has been replace by a more specific symantec_endpoint
package that is not generated/derived from RSA Netwitness parsers.

This change increments the patch version and updates the title, description,
and readme to indicate that it is deprecated.

See https://github.com/elastic/integrations/pull/2187 for the replacement.